diploms_generator/pyproject.toml
[tool.poetry]
name = "diploms"
version = "0.1.0"
description = ""
authors = ["Zahar Chernenko <zaharchernenko35@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pylint = "^3.3.1"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
